Notifikasi Daftar Pengiriman barang untuck Markoantar & Opsi Markoantar untuk Konfirmasi dari Per Item menjadi Per Order / PO

+++ dashboard.php
@@ -20,7 +20,9 @@
 	if($_GET["delete_store_sales_list"] == 1 && $_GET["po_no"] != ""){
 		$po_no = $_GET["po_no"];
 		$transaction_status = $db->fetch_single_data("transactions","status",["po_no" => $po_no]);
-		if($transaction_status < 3){
+		$transaction_cart_group = $db->fetch_single_data("transactions","cart_group",["po_no" => $po_no]);
+		$is_cod = $db->fetch_single_data("transaction_payments","id",["cart_group" => $transaction_cart_group,"payment_type_id" => "-1"]);
+		if($is_cod && $transaction_status <= 3){
 			$db->addtable("transaction_details");	$db->where("transaction_id","(SELECT id FROM transactions WHERE po_no='".$po_no."' AND seller_user_id='".$__user_id."')","s","IN");	$db->delete_();
 			$db->addtable("transactions");			$db->where("po_no",$po_no);	$db->where("seller_user_id",$__user_id);	$db->delete_();
 			$_SESSION["message"] = v("delete_transaction_by_po_no_success");
@@ -147,7 +149,7 @@
 			<?php if($__forwarder_id > 0){ ?>
 			<div id="panel_list_of_delivering_goods" class="panel panel-default">
 				<div class="panel-heading" data-toggle="collapse" data-parent="#dashboard" href="#list_of_delivering_goods" onclick="changeState('list_of_delivering_goods');">
-					<h3 class="panel-title"><img src="assets/sent.png" height="20"> <b><?=v("list_of_delivering_goods");?></b><span class="notification-counter" style="visibility:hidden;" id="notifStoreSalesListTabCount"></span></h3>
+					<h3 class="panel-title"><img src="assets/sent.png" height="20"> <b><?=v("list_of_delivering_goods");?></b><span class="notification-counter" style="visibility:hidden;" id="notifDeliveringGoodsTabCount"></span></h3>
 				</div>
 				<div id="list_of_delivering_goods" class="panel-collapse collapse <?=($_GET["tabActive"] == "list_of_delivering_goods")?"in":"";?>"><div class="panel-body"><?php include_once "dashboard_list_of_delivering_goods.php";?></div></div>
 			</div><br>
